+++ feature4/feature4/src/Components/Connect/Connect.js
@@ -1,33 +1,29 @@
 import React, { useState, useEffect } from "react";
 import ReqList from "../Main/ReqList.js";
 import { getAllRequests, Requests} from "../../Services/RequestService";
-
+import CourseForm from "./CourseForm.js";
   const Connect = () => {
+    const [reqs, setReqs] = useState([]);
+    const [courses, setCourses] = useState([]);
+    const [user, setUser] = useState({ canHelp: [], needHelp: [] });
+
     // const data = useFetch("https://jsonplaceholder.typicode.com/todos/");
      // console.log("data: ", data);
      // Variables in the state to hold data
-     const [reqs, setReqs] = useState([]);
-     // UseEffect to run when the page loads to
-     // obtain async data and render
-     useEffect(() => {
-       if (Requests.collection.length) {
-         setReqs(Requests.collection);
-       } else {
-         getAllRequests().then((reqs) => {
-           console.log(reqs);
-           setReqs(reqs);
-         });
-       }
-     }, []);
+
+      const onChange = (_e) => {
+        console.log('onChange called');
+      }
    
   // on this page students can connect and chat with other users who they matched with
     return (
       <section>
         <div className="centerText">
         <h1> Connect with other students </h1>
-        <p> This is the connect component </p>
+        <p> Fill out the forms below so we can connect you with other students that can help you and you can help! </p>
         </div>
         <div className="db-list">
+        <CourseForm courses={courses} user={user} setUser={setUser} onChange={onChange}/>
         <ReqList reqs={reqs} />
         </div>