+++ feature4/feature4/src/Components/Header/Header.js
@@ -3,35 +3,48 @@ import { Link } from "react-router-dom";
 import { checkUser, logoutUser } from "../Auth/AuthService";
 
 const Header = () => {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-
-  useEffect(() => {
-    const userAuthenticated = checkUser();
-    setIsLoggedIn(userAuthenticated);
-  }, []);
-
-  useEffect(() => {
-    const userAuthenticated = checkUser();
-    setIsLoggedIn(userAuthenticated);
-  }, [isLoggedIn]);
+  const [isLoggedIn, setIsLoggedIn] = useState(checkUser());
 
   const handleLogout = () => {
     logoutUser()
-    .then(() => {
-      setIsLoggedIn(false);
-    })
-    .catch((error) => {
-      // Handle any error that occurs during logout
-      console.error("Logout Error: ", error);
-    });
+      .then(() => {
+        setIsLoggedIn(checkUser());
+      })
+      .catch((error) => {
+        console.error("Logout Error: ", error);
+      });
   };
+
+  useEffect(() => {
+    // This will constantly check the user status every second
+    const intervalID = setInterval(() => setIsLoggedIn(checkUser()), 1000);
+
+    // This function will run when component unmounts. Cleaning up the interval.
+    return () => clearInterval(intervalID);
+  }, []);
 
   return (
     <header>
       <div className="nav-bar-container">
         <nav>
-          <ul className="navigation">
-            <li>
+            {isLoggedIn ? (
+              <ul className="navigation">
+                <li>
+              <Link to="/">Home</Link>
+            </li>
+                   <li>
+                    <Link to="/connect">Connect</Link>
+                  </li>
+                  <li>
+                    <Link to="/yourmatches">Your Matches</Link>
+                  </li>
+                  <li>
+                    <button type="submit" className="btn btn-primary" onClick={handleLogout}>Logout</button>
+                  </li>
+                  </ul>
+            ) : 
+            <ul className="navigation">
+              <li>
               <Link to="/">Home</Link>
             </li>
             <li>
@@ -40,19 +53,9 @@ const Header = () => {
             <li>
               <Link to="/auth/login">Login</Link>
             </li>
-            
-            <li>
-              <Link to="/connect">Connect</Link>
-            </li>
-            <li>
-              <Link to="/yourmatches">Your Matches</Link>
-            </li>
-            {isLoggedIn ? (
-                  <li>
-                    <button type="submit" className="btn btn-primary" onClick={handleLogout}>Logout</button>
-                  </li>
-            ) : null}
-          </ul>
+            </ul>
+            }
+        
         </nav>
       </div>
     </header>